🚀 Add react-leaflet-markercluster package and display markers on the map
The code changes include adding the "react-leaflet-markercluster" package to the project's dependencies in the package.json file. This package allows for clustering of markers on the map.

In the Home.jsx file, two new plant objects are added to the "plantes" array with their respective latitude and longitude coordinates. These objects represent markers that will be displayed on the map.

Additionally, a Grid item is modified to adjust the layout of components on the page. The MapContainer component now renders multiple Marker components based on the "plantes" array, each displaying a popup with the name of the plant.

ℹ️ Note: The MarkerClusterGroup component is commented out for now.